Alert users about high gas fees
When a spike in gas fees occurs, such as during an NFT drop, we want to
not only inform the user about this, but also dissuade them from using a
higher gas fee (as we have proved in testing that high gas fees can
cause bidding wars and exacerbate the situation).

To gauge whether there is a spike, or whether the network is congested
in general, we can look at recent blocks on the network to judge whether
users are paying more fees than normal. We already estimate gas fees
on a 15-second cadence when users are editing the gas fees for a new
transaction or swap, so for EIP-1559-capable networks, we can simply add
an additional request to that cadence. This capability is provided by
GasFeeController, so that's what does the real work. This commit simply
wires things up.